move error handling and types of our main component
